Fix up req and design decision doc build (tmp).
Building this doc should go through the cmake build,
but OFF by default. Do not fail cmake if pdflatex etc not found.

Currently also requires inkscape for svg->pdf. This dep has to go,
but I don't like to put a pdf into version control.
(.svg is a text fmt and also questionable; .dia is binary, so also poor.)

to be continued...