Switch SOVERSION from MAJOR to MAJOR.MINOR
I found out that linking by default to .so.2 is not so useful,
because if the size of passed data structures changes, the dyn linker
correctly refuses to link and users need to rebuild anyway.
As I already knew, MAJOR.MINOR is really our API version, so do that.

In the end, it matters little, because almost all users will have to
rebuild on every release anyway. Also because of template code in .h/.tcc.