Add tbb and generic header info util. Also rename py utils '-' -> '_'.
You (almost) cannot import a py module with a '-' in it, so rename all
our py utils to have '_', because we really want to be able to reuse
them, like with the generic header info util. It figures out the
FILETYPE and then dispatches to the bf or tbb variant.

Also added to the tests.